Add a `Myself` abstraction to the Butterfly server
Previously, we just stored a bare `Member` to represent a Butterfly
server's identity within a Habitat network. When we needed to
increment "our" incarnation value, we would just manipulate the
`Member` directly. Adding an `increment_incarnation` function would
help to abstract this logic, but would potentially encourage anyone to
increment an incarnation number (we still need to refactor `Member`
such that `id` at least isn't a `pub` field).

By creating a new type `Myself` that wraps a `Member`, we are
expressing a stronger constraint ("This Member is special because it
is *me*") which then allows us to add the desired increment method
cleanly.

A future commit will take further advantage of this by folding
persistence of an incarnation number into this increment method.

Signed-off-by: Christopher Maier <cmaier@chef.io>